requirements.txt
google-cloud-storage==3.7.0